src/requirements.txt
ultralytics==8.0.190 # Chỉ định phiên bản để đảm bảo tính ổn định
opencv-python-headless
numpy
pyyaml
tqdm